+++ sdk/issuingTokenActivation/index.js
@@ -1,0 +1,4 @@
+const issuingTokenActivation = require('./issuingTokenActivation.js');
+
+exports.parse = issuingTokenActivation.parse;
+exports.IssuingTokenActivation = issuingTokenActivation.IssuingTokenActivation;