+++ sdk/issuingTokenDesign/index.js
@@ -1,0 +1,7 @@
+const issuingTokenDesign = require('./issuingTokenDesign.js');
+
+exports.get = issuingTokenDesign.get;
+exports.query = issuingTokenDesign.query;
+exports.page = issuingTokenDesign.page;
+exports.pdf = issuingTokenDesign.pdf;
+exports.IssuingTokenDesign = issuingTokenDesign.IssuingTokenDesign;